refactor(dg): create filterStore
