+++ packages/data-grid/src/stores/index.ts
@@ -1,2 +1,3 @@
 export * from './data'
 export * from './pagination'
+export * from './filter'
